♻️ refactor: useAlarmFlow 구조 개선
