+++ Makcha/src/pages/Alarm/apis/alerts.ts
@@ -1,4 +1,5 @@
 import { api } from "../../../apis/api";
+import type { CandidateStep } from "../types/candidate";
 
 type BaseResponse<T> = {
     successCode: string;
@@ -8,8 +9,8 @@ type BaseResponse<T> = {
 };
 
 export type CreateAlertBody = {
-    cacheKey: string;   
-    alert_time?: number;  
+    cacheKey: string;
+    alert_time?: number;
 };
 
 export type CreateAlertResult = {
@@ -18,5 +19,31 @@ export type CreateAlertResult = {
 
 export async function createAlert(body: CreateAlertBody) {
     const res = await api.post<BaseResponse<CreateAlertResult>>("/api/alerts", body);
+    return res.data.result;
+}
+
+export type AlertDetailResult = {
+    notification_id: number;
+
+    is_optimal: boolean;
+    lines: string[];
+    total_duration_min: number;
+    transfer_count: number;
+    walking_time_min: number;
+    minutes_left: number;
+
+    departure_at?: string;
+    arrival_at?: string;
+
+    route_id?: string;
+    route_token: string;
+
+    steps: CandidateStep[];
+};
+
+export async function fetchAlertDetail(notificationId: number) {
+    const res = await api.get<BaseResponse<AlertDetailResult>>(
+        `/api/alerts/${notificationId}/detail`
+    );
     return res.data.result;
 }
